requirements.txt
langchain
newsdataapi
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2